Added Uber Lyra
This version of Lyra has the capability of manipulating peers, nodes, certificate authorities and orderer's through code. This means the testing phase can begin once this is finished.
The available methods allow to creating users, creating channels, joining channels, installing chaincode, instantiating chaincode, invoking the ledger, querying the ledger, checking a single block, checking a transaction and so on.
The amount of code poured into this means that the CLI must go as it cant be developed flexible enough to support the commands, but visual elements will remain